requirements.txt
Django==3.2.7
djangorestframework
markdown
django-filter
django-colorfield
pillow
djangorestframework-simplejwt
djoser
reportlab
pytest
flake8